+++ src/ck-core/ckfutures.h
@@ -3,6 +3,11 @@
 
 #include "CkFutures.decl.h"
 
+#ifdef __cplusplus
+extern CkMarshallMsg *CkAllocateMarshallMsg(int size, const CkEntryOptions *opts);
+#include <type_traits>
+#endif
+
 /**
 \addtogroup CkFutures
 \brief Futures--ways to block Converse threads on remote events.
@@ -10,7 +15,7 @@
 These routines are implemented in ckfutures.C.
 */
 /*@{*/
-typedef int CkFutureID;
+typedef unsigned long long CkFutureID;
 typedef struct _CkFuture {
   CkFutureID id;
   int        pe;
@@ -53,6 +58,38 @@ void *CkWaitReleaseFuture(CkFutureID futNum);
 void _futuresModuleInit(void);
 
 #ifdef __cplusplus
+}
+
+namespace ck {
+  template<typename T>
+  class future {
+    CkFuture handle_;
+public:
+    future() { handle_ = CkCreateFuture(); }
+    future(const future<T>& other) { handle_ = other.handle_; }
+
+    T get() {
+      CkMarshallMsg *msg = (CkMarshallMsg *)CkWaitFuture(handle_);
+      PUP::fromMem p(msg->msgBuf);
+      PUP::detail::TemporaryObjectHolder<T> holder;
+      p | holder;
+      delete msg;
+      return std::move(holder.t);
+    }
+
+    void set(const T& value) {
+      PUP::sizer s;
+      s | (typename std::decay<decltype(value)>::type&) value;
+      CkMarshallMsg *msg = CkAllocateMarshallMsg(s.size(), NULL);
+      PUP::toMem p((void *)msg->msgBuf);
+      p | (typename std::decay<decltype(value)>::type&) value;
+      CkSendToFuture(handle_, msg);
+    }
+
+    bool probe() { return CkProbeFuture(handle_); }
+    void release() { CkReleaseFuture(handle_); }
+    void pup(PUP::er& p){ p | handle_; }
+  };
 }
 #endif
